requirements.txt
plexapi
pyyaml
